refactor: remove mcp-renderer component
